[DEC-2081] Audit app/prepare/other_msgs.go and app/prepare/transactions.go (#554)
* [DEC-2081] Audit app/prepare/other_msgs.go and app/prepare/transactions.go

Mainly swap to use uint64 for value that is meant to never be negative which will prevent `<0` cases during coding.
Note that the testing of what is and isn't allowed should be covered by lib/ante/disallow_msg.go audit.

+++ protocol/app/prepare/prepare_proposal.go
@@ -114,7 +114,7 @@ func PrepareProposalHandler(
 		}
 
 		// Gather "Other" group messages.
-		otherBytesAllocated := txs.GetAvailableBytes() / int64(4) // ~25% of the remainder.
+		otherBytesAllocated := txs.GetAvailableBytes() / 4 // ~25% of the remainder.
 		// filter out txs that have disallow messages.
 		txsWithoutDisallowMsgs := RemoveDisallowMsgs(ctx, txConfig.TxDecoder(), req.Txs)
 		otherTxsToInclude, otherTxsRemainder := GetGroupMsgOther(txsWithoutDisallowMsgs, otherBytesAllocated)
